2026-03-10-python-chatbot-tools/demo/requirements.txt
# 파이썬 챗봇 프레임워크 데모 의존성
# Python 3.10+ 권장

# ─── LLM API 클라이언트 ───────────────────────────────
anthropic>=0.40.0       # Anthropic Claude API 클라이언트

# ─── 환경변수 관리 ───────────────────────────────────
python-dotenv>=1.0.0    # .env 파일 로드

# ─── Streamlit ────────────────────────────────────────
streamlit>=1.35.0       # Streamlit 프레임워크

# ─── Gradio ──────────────────────────────────────────
gradio>=4.36.0          # Gradio 프레임워크

# ─── Chainlit ────────────────────────────────────────
chainlit>=1.1.0         # Chainlit 프레임워크

# ─── Reflex ──────────────────────────────────────────
reflex>=0.5.0           # Reflex 프레임워크

# ─── Panel ───────────────────────────────────────────
panel>=1.4.0            # Panel 프레임워크
bokeh>=3.4.0            # Panel 의존성

# ─── Mesop ───────────────────────────────────────────
mesop>=0.10.0           # Google Mesop 프레임워크

# ─── Voilà ───────────────────────────────────────────
voila>=0.5.0            # Voilà 프레임워크
ipywidgets>=8.0.0       # Jupyter 위젯
jupyter>=1.0.0          # Jupyter 노트북
